new class TokenRequestValidator added
